fix: alpha values are in one format for Hex and another for RGB
The problem with Alpha - Alpha values in hex run from 0 - 255 but rgba takes 0 - 1 (a percentage).  This leaves us with a little problem. We have to do a bit of math to get our Hex values for Alpha. Moving Alpha from decimal to hex requires rounding and means the original values are destroyed.

I've tried to account for the differences, but users will just have to know that if they want to use Alpha this library isn't 100 percent accurate right now.

Future features will require better planning, so we can store Alpha in the exact format it was given to us rather than the one that converts to hex.