feat: switch core to use vite build
We're switching to vite build to take advantage of some built in sass tools and dropping emotion. This was a hard decision for me, but ultimately I hated how every new project required a set of changes just to get emotion running. It was our heaviest dependency by basis of effort.